+++ src/Mzinga.Viewer/App.axaml.cs
@@ -25,12 +25,13 @@
 // THE SOFTWARE.
 
 using System;
+using System.ComponentModel;
 using System.IO;
-using System.Reflection;
 
 using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
+using Avalonia.Styling;
 
 using Mzinga.Engine;
 
@@ -53,9 +54,14 @@ namespace Mzinga.Viewer
 
         public string ViewerConfigPath { get; private set; }
 
+        public static IStyle FluentLight { get; private set; }
+        public static IStyle FluentDark { get; private set; }
+
         public override void Initialize()
         {
             AvaloniaXamlLoader.Load(this);
+            FluentLight = AvaloniaXamlLoader.Load(new Uri("avares://Avalonia.Themes.Fluent/FluentLight.xaml")) as IStyle;
+            FluentDark = AvaloniaXamlLoader.Load(new Uri("avares://Avalonia.Themes.Fluent/FluentDark.xaml")) as IStyle;
         }
 
         public override void OnFrameworkInitializationCompleted()
@@ -92,6 +98,8 @@ namespace Mzinga.Viewer
 
             AppViewModel.Init(parameters);
             DataContext = AppVM;
+
+            Current.Styles[0] = AppVM.ViewerConfig.VisualTheme == VisualTheme.Dark ? FluentDark : FluentLight;
 
             if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
             {